Suporte a divisão do index por tenant

+++ nsj_embedding_search/dao.py
@@ -17,23 +17,26 @@ class DAO:
         self,
         metadata: dict = None,
         limit_results: int = 1000,
+        tenant: int = 0,
     ):
 
-        where_clause = ""
+        where_clause = "WHERE tenant = %(tenant)s"
         if metadata is not None:
-            where_clause = """
-            WHERE metadata @> %(metadata)s
+            where_clause += """
+            and metadata @> %(metadata)s
             """
 
         sql = f"""
-        SELECT external_id, title, embedding, content, reference, metadata, chunck_number, total_chunks, created_at, updated_at
+        SELECT
+            external_id, title, embedding, content, reference,
+            metadata, chunck_number, total_chunks, created_at, updated_at, tenant
         FROM {self._index_table}
         {where_clause}
         ORDER BY external_id, chunck_number, total_chunks
         LIMIT {limit_results}
         """
 
-        return self._db.execute(sql, metadata=json.dumps(metadata))
+        return self._db.execute(sql, metadata=json.dumps(metadata), tenant=tenant)
 
     def insert(
         self,
@@ -45,11 +48,12 @@ class DAO:
         metadata: dict[str, any],
         chunck_number: int,
         total_chunks: int,
+        tenant: int = 0,
     ):
         sql = f"""
         insert into {self._index_table}
-        (external_id, title, embedding, content, reference, metadata, chunck_number, total_chunks)
-        values (%(external_id)s, %(title)s, %(embedding)s, %(content)s, %(reference)s, %(metadata)s, %(chunck_number)s, %(total_chunks)s)
+        (external_id, title, embedding, content, reference, metadata, chunck_number, total_chunks, tenant)
+        values (%(external_id)s, %(title)s, %(embedding)s, %(content)s, %(reference)s, %(metadata)s, %(chunck_number)s, %(total_chunks)s, %(tenant)s)
         """
 
         self._db.execute(
@@ -62,12 +66,14 @@ class DAO:
             metadata=json.dumps(metadata),
             chunck_number=chunck_number,
             total_chunks=total_chunks,
+            tenant=tenant,
         )
 
-    def delete(self, external_id: uuid.UUID):
+    def delete(self, external_id: uuid.UUID, tenant: int = 0):
         sql = f"""
         DELETE FROM {self._index_table}
         WHERE external_id = %(external_id)s
+        and tenant = %(tenant)s
         """
 
-        self._db.execute(sql, external_id=external_id)
+        self._db.execute(sql, external_id=external_id, tenant=tenant)
